requirements.txt
asgiref==3.2.10
boto3==1.16.36
botocore==1.19.36
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-active-link==0.1.6
django-appconf==1.0.4
django-crispy-forms==1.9.1
django-environ==0.4.5
django-filter==2.3.0
django-image-renderer==0.1.3
django-imagekit==4.0.2
django-json-field==0.5.7
django-material==1.4.3
django-picklefield==2.1.1

django-social-share==2.1.0
django-storages==1.9.1
django-taggit==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.10
drf-nested-routers==0.91
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.7.0
jmespath==0.9.5
Markdown==3.2.2
mock==4.0.2
pilkit==2.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.3.1
urllib3==1.25.9
uuid==1.30
whitenoise==5.0.1
zipp==3.1.0
